Presampling + adaptive algorithms created.
Ideally the presampling is a layer on top of the adaptive rather than
having that code duplicated there. And ideally the adaptive can
remove chunks at random offsets rather than scanning purely left to
right, which is vulnerable to statistical bias.

And the chunk phase, both in the adaptive and the presampling versions,
will attempt to remove characters that were already found to be
important, which could be fixed.

Results on fake data are essentially *exactly* what I was hoping they
would be. Surprisingly close to optimal in extreme cases (ok, over 3x,
but still), better than delta debugging even in the extreme cases (99.5%
chance of waste, 99.5% chance waste repeats itself), and nearly optimal
in the worst cases (~505 tests for most of the >50% side, where the 5
likely come from the 5 presamples we do since we don't prevent checking
them again in the chunking phase).

Ready to try to wire up to real world tests, I think.