Competing planners to build better trees at various tree sizes.
Tree builder combinators, entropy builder, planner builder.

For decision trees of size 10 or less, use the optimal builder.

For decision trees of size 100 or less, use the entropy/huffman/optimal
combinator.

For decision trees of other sizes, cap at 1000 and use the huffmanlike
builder.

To be fair...this is a lot of additional complexity, for little
benchmarkable gain.