New algorithm, "Sliding window" approach, for mostly benchmarking.
This approach opens a sliding window into the input, and enumerates all
possible simplification stream forms of that window. It then enumerates
all tests, and picks the test with the most information gain. As the
product of tests & states fall below a threshold, it expands the window.

This algorithm does not (yet....it *could*) discover/maintain/update a
markov model but rather requires one pre-fed, which makes it currently
not useful for more than benchmarking. Furthermore, it is slower than
than runlength-based entropy debugging, even though it currently
runs slightly fewer tests especially for instance in scenarios of
anticorrelation in the markov model.

For benchmarking, it is revealing because it does not seem to have
significantly different performance characteristics than entropy
debugging aside from anti-correlated cases. This seems to indicate that
deviation from currently calculated lower bounds in most scenarios is
not necessarily resulting from flaws in algorithm design, but rather
flaws in the lower bound itself.