Add an A* search for optimal trees -- but, currently subpar.
While it does produce optimal trees (and suboptimal ones with less
computation), it does not do so in less time than the simpler optimal
algorithm.

This is probably almost entirely attributable to the fact that common
subproblems frequently must be resolved (ie, what's the optimal
arrangement of decisions 4-12) across nodes in the graph. An addition
that can cache the results of these subproblems is not incredibly
obvious in terms of fitting into the A* algorithm, but, if that is added
then this could become the fastest way to generate optimal trees.