Remove naive entropy builder.
The whole algorithm was based on a misunderstanding. The goal of the
naive entropy builder should not be to split entropy, but to maximize
information gain. In this case, IG = H(X) - H(Y). Thefore, we just want
to get a close to 50/50 test -- that's the even probability builder.

Actually splitting on entropy is bad. We may have a leftmost leaf with
50% probability, and then hundreds of unlikely things after. The entropy
of the 50% case is small, and the entropy of the set of unlikely things
is large. Therefore, splitting the entropy means putting unlikely events
in a bucket with the 50% case. This is all around bad.

Deleted.